Add support dynamic workspaces. (#33)
Updated the supported Bazel versions.
Changed the --workspace flag for test runners. Previously, it was the path to the child's WORKSPACE file. Now, it is the path to the workspace directory.
Updated the default test runner and the custom test runners in the repository for the new --workspace flag meaning.
Updated bazel_integration_test to allow declarations without workspace_path.