fix: hidden files being copied to the wrong directory (#404)
The initial copy will copy over hidden files in sub-directories. The
copy `find` logic for hidden files was looking through the entire tree.
We really just want to copy the hidden files in the top-level directory.

Also, adjusted the glob pattern for the "non-hidden" file copy per a
shellcheck warning.