fix: do not register a python toolchain (#523)
Do not register a python toolchain: this is unnecessary since
rules_python registers a default toolchain already. Moreover, this
toolchain registration could negatively affect a user of
`rules_bazel_integration_test` by registering a toolchain the root
module may not want to use (and that might be selected depending on that
module's dependency ordering).

Fixes #520